refactor(tests) Tests to fixtures - propertyLiterals (#760)
[skip publish]
