+++ packages/babel-plugin-transform-property-literals/__tests__/fixtures/computed-property/expected.js
@@ -1,0 +1,5 @@
+({
+  [a]: null,
+  [ಠ_ಠ]: "foo",
+  ಠ_ಠ: "bar"
+});
