+++ packages/babel-plugin-transform-property-literals/__tests__/fixtures/es5-unicode-property/expected.js
@@ -1,0 +1,9 @@
+// FIXME: The test name states that the property names should be transformed,
+// but the `12e34` one isn’t transformed in the test
+
+// prettier-ignore
+({
+  ಠ_ಠ: "bar",
+  "12e34": "wut",
+  Ǽ: "AE"
+});
