+++ packages/babel-plugin-transform-property-literals/__tests__/fixtures/invalid-es5-property/actual.js
@@ -1,0 +1,5 @@
+// prettier-ignore
+({
+  "\u2118": "wp",
+  "𐊧": "foo"
+});
