+++ packages/babel-plugin-transform-property-literals/__tests__/fixtures/invalid-es5-property/expected.js
@@ -1,0 +1,5 @@
+// prettier-ignore
+({
+  "℘": "wp",
+  "𐊧": "foo"
+});
