+++ packages/babel-plugin-transform-property-literals/__tests__/fixtures/invalid-identifiers/expected.js
@@ -1,0 +1,7 @@
+// FIXME: The test name states that no transformation should take place,
+// but the actual test specifies that the quotes should be stripped.
+//prettier-ignore
+({
+  default: null,
+  import: null
+});
