+++ packages/babel-plugin-transform-property-literals/__tests__/fixtures/leading-zeros/actual.js
@@ -1,0 +1,4 @@
+var data = {
+  "00": 1,
+  "01": 2
+};
