+++ packages/babel-plugin-transform-property-literals/__tests__/fixtures/leading-zeros/expected.js
@@ -1,0 +1,4 @@
+var data = {
+  "00": 1,
+  "01": 2
+};
