+++ packages/babel-plugin-transform-property-literals/__tests__/fixtures/strip-quotes/expected.js
@@ -1,0 +1,2 @@
+// prettier-ignore
+var x = { foo: "bar", 1: "baz", test: null };
